@@ -69,7 +69,7 @@
     "start": "vite --host",
     "build": "vite build",
     "coverage": "VITE_COVERAGE=true vite build",
-    "lint": "eslint --ext .js,.jsx,.ts,.tsx --fix .",
+    "lint": "eslint --format visualstudio --ext .js,.jsx,.ts,.tsx --fix .",
     "test": "vitest run --coverage",
     "test:watch": "vitest --coverage",
     "test:preview": "vitest-preview",
@@ -108,13 +108,12 @@
     "@vitejs/plugin-react": "5.0.1",
     "@vitest/coverage-istanbul": "3.2.4",
     "esbuild": "0.25.9",
-    "eslint": "8.57.1",
+    "eslint": "9.32.0",
     "eslint-config-prettier": "10.1.8",
-    "eslint-config-react-app": "7.0.1",
     "eslint-formatter-rdjson": "1.0.6",
+    "eslint-formatter-visualstudio": "8.40.0",
     "eslint-import-resolver-typescript": "4.4.4",
     "eslint-plugin-import": "2.32.0",
-    "eslint-plugin-jsx-a11y": "6.10.2",
     "eslint-plugin-prettier": "5.5.4",
     "eslint-plugin-react": "7.37.5",
     "eslint-plugin-react-hooks": "5.2.0",
@@ -123,7 +122,7 @@
     "prettier": "3.6.2",
     "typescript": "5.9.2",
     "vite": "7.1.3",
-    "vite-plugin-checker": "0.8.0",
+    "vite-plugin-checker": "0.10.2",
     "vite-plugin-istanbul": "7.1.0",
     "vite-plugin-svgr": "4.5.0",
     "vite-tsconfig-paths": "5.1.4",
