@@ -1,4 +1,4 @@
-module.exports = {
+export default {
     extends: ["@commitlint/config-conventional"],
     rules: {
         "body-max-line-length": [2, "always", "Infinity"],
